TODO : On click get posX/Y and at it to GameBoard
